Remove deprecated GetServers RPC
