+++ gen/proto/ts/teleport/lib/teleterm/v1/service_pb.ts
@@ -826,56 +826,6 @@ export interface SetGatewayLocalPortRequest {
     localPort: string;
 }
 /**
- * @generated from protobuf message teleport.lib.teleterm.v1.GetServersRequest
- */
-export interface GetServersRequest {
-    /**
-     * @generated from protobuf field: string cluster_uri = 1;
-     */
-    clusterUri: string;
-    /**
-     * @generated from protobuf field: int32 limit = 2;
-     */
-    limit: number;
-    /**
-     * @generated from protobuf field: string start_key = 3;
-     */
-    startKey: string;
-    /**
-     * @generated from protobuf field: string search = 4;
-     */
-    search: string;
-    /**
-     * @generated from protobuf field: string query = 5;
-     */
-    query: string;
-    /**
-     * @generated from protobuf field: string sort_by = 6;
-     */
-    sortBy: string;
-    /**
-     * @generated from protobuf field: string search_as_roles = 7;
-     */
-    searchAsRoles: string;
-}
-/**
- * @generated from protobuf message teleport.lib.teleterm.v1.GetServersResponse
- */
-export interface GetServersResponse {
-    /**
-     * @generated from protobuf field: repeated teleport.lib.teleterm.v1.Server agents = 1;
-     */
-    agents: Server[];
-    /**
-     * @generated from protobuf field: int32 total_count = 2;
-     */
-    totalCount: number;
-    /**
-     * @generated from protobuf field: string start_key = 3;
-     */
-    startKey: string;
-}
-/**
  * @generated from protobuf message teleport.lib.teleterm.v1.GetAuthSettingsRequest
  */
 export interface GetAuthSettingsRequest {
@@ -4106,164 +4056,6 @@ class SetGatewayLocalPortRequest$Type extends MessageType<SetGatewayLocalPortReq
  */
 export const SetGatewayLocalPortRequest = new SetGatewayLocalPortRequest$Type();
 // @generated message type with reflection information, may provide speed optimized methods
-class GetServersRequest$Type extends MessageType<GetServersRequest> {
-    constructor() {
-        super("teleport.lib.teleterm.v1.GetServersRequest", [
-            { no: 1, name: "cluster_uri", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
-            { no: 2, name: "limit", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
-            { no: 3, name: "start_key", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
-            { no: 4, name: "search", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
-            { no: 5, name: "query", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
-            { no: 6, name: "sort_by", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
-            { no: 7, name: "search_as_roles", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
-        ]);
-    }
-    create(value?: PartialMessage<GetServersRequest>): GetServersRequest {
-        const message = globalThis.Object.create((this.messagePrototype!));
-        message.clusterUri = "";
-        message.limit = 0;
-        message.startKey = "";
-        message.search = "";
-        message.query = "";
-        message.sortBy = "";
-        message.searchAsRoles = "";
-        if (value !== undefined)
-            reflectionMergePartial<GetServersRequest>(this, message, value);
-        return message;
-    }
-    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: GetServersRequest): GetServersRequest {
-        let message = target ?? this.create(), end = reader.pos + length;
-        while (reader.pos < end) {
-            let [fieldNo, wireType] = reader.tag();
-            switch (fieldNo) {
-                case /* string cluster_uri */ 1:
-                    message.clusterUri = reader.string();
-                    break;
-                case /* int32 limit */ 2:
-                    message.limit = reader.int32();
-                    break;
-                case /* string start_key */ 3:
-                    message.startKey = reader.string();
-                    break;
-                case /* string search */ 4:
-                    message.search = reader.string();
-                    break;
-                case /* string query */ 5:
-                    message.query = reader.string();
-                    break;
-                case /* string sort_by */ 6:
-                    message.sortBy = reader.string();
-                    break;
-                case /* string search_as_roles */ 7:
-                    message.searchAsRoles = reader.string();
-                    break;
-                default:
-                    let u = options.readUnknownField;
-                    if (u === "throw")
-                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
-                    let d = reader.skip(wireType);
-                    if (u !== false)
-                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
-            }
-        }
-        return message;
-    }
-    internalBinaryWrite(message: GetServersRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
-        /* string cluster_uri = 1; */
-        if (message.clusterUri !== "")
-            writer.tag(1, WireType.LengthDelimited).string(message.clusterUri);
-        /* int32 limit = 2; */
-        if (message.limit !== 0)
-            writer.tag(2, WireType.Varint).int32(message.limit);
-        /* string start_key = 3; */
-        if (message.startKey !== "")
-            writer.tag(3, WireType.LengthDelimited).string(message.startKey);
-        /* string search = 4; */
-        if (message.search !== "")
-            writer.tag(4, WireType.LengthDelimited).string(message.search);
-        /* string query = 5; */
-        if (message.query !== "")
-            writer.tag(5, WireType.LengthDelimited).string(message.query);
-        /* string sort_by = 6; */
-        if (message.sortBy !== "")
-            writer.tag(6, WireType.LengthDelimited).string(message.sortBy);
-        /* string search_as_roles = 7; */
-        if (message.searchAsRoles !== "")
-            writer.tag(7, WireType.LengthDelimited).string(message.searchAsRoles);
-        let u = options.writeUnknownFields;
-        if (u !== false)
-            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
-        return writer;
-    }
-}
-/**
- * @generated MessageType for protobuf message teleport.lib.teleterm.v1.GetServersRequest
- */
-export const GetServersRequest = new GetServersRequest$Type();
-// @generated message type with reflection information, may provide speed optimized methods
-class GetServersResponse$Type extends MessageType<GetServersResponse> {
-    constructor() {
-        super("teleport.lib.teleterm.v1.GetServersResponse", [
-            { no: 1, name: "agents", kind: "message", repeat: 1 /*RepeatType.PACKED*/, T: () => Server },
-            { no: 2, name: "total_count", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
-            { no: 3, name: "start_key", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
-        ]);
-    }
-    create(value?: PartialMessage<GetServersResponse>): GetServersResponse {
-        const message = globalThis.Object.create((this.messagePrototype!));
-        message.agents = [];
-        message.totalCount = 0;
-        message.startKey = "";
-        if (value !== undefined)
-            reflectionMergePartial<GetServersResponse>(this, message, value);
-        return message;
-    }
-    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: GetServersResponse): GetServersResponse {
-        let message = target ?? this.create(), end = reader.pos + length;
-        while (reader.pos < end) {
-            let [fieldNo, wireType] = reader.tag();
-            switch (fieldNo) {
-                case /* repeated teleport.lib.teleterm.v1.Server agents */ 1:
-                    message.agents.push(Server.internalBinaryRead(reader, reader.uint32(), options));
-                    break;
-                case /* int32 total_count */ 2:
-                    message.totalCount = reader.int32();
-                    break;
-                case /* string start_key */ 3:
-                    message.startKey = reader.string();
-                    break;
-                default:
-                    let u = options.readUnknownField;
-                    if (u === "throw")
-                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
-                    let d = reader.skip(wireType);
-                    if (u !== false)
-                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
-            }
-        }
-        return message;
-    }
-    internalBinaryWrite(message: GetServersResponse, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
-        /* repeated teleport.lib.teleterm.v1.Server agents = 1; */
-        for (let i = 0; i < message.agents.length; i++)
-            Server.internalBinaryWrite(message.agents[i], writer.tag(1, WireType.LengthDelimited).fork(), options).join();
-        /* int32 total_count = 2; */
-        if (message.totalCount !== 0)
-            writer.tag(2, WireType.Varint).int32(message.totalCount);
-        /* string start_key = 3; */
-        if (message.startKey !== "")
-            writer.tag(3, WireType.LengthDelimited).string(message.startKey);
-        let u = options.writeUnknownFields;
-        if (u !== false)
-            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
-        return writer;
-    }
-}
-/**
- * @generated MessageType for protobuf message teleport.lib.teleterm.v1.GetServersResponse
- */
-export const GetServersResponse = new GetServersResponse$Type();
-// @generated message type with reflection information, may provide speed optimized methods
 class GetAuthSettingsRequest$Type extends MessageType<GetAuthSettingsRequest> {
     constructor() {
         super("teleport.lib.teleterm.v1.GetAuthSettingsRequest", [
@@ -5936,7 +5728,6 @@ export const TerminalService = new ServiceType("teleport.lib.teleterm.v1.Termina
     { name: "StartHeadlessWatcher", options: {}, I: StartHeadlessWatcherRequest, O: StartHeadlessWatcherResponse },
     { name: "ListDatabaseUsers", options: {}, I: ListDatabaseUsersRequest, O: ListDatabaseUsersResponse },
     { name: "ListDatabaseServers", options: {}, I: ListDatabaseServersRequest, O: ListDatabaseServersResponse },
-    { name: "GetServers", options: {}, I: GetServersRequest, O: GetServersResponse },
     { name: "GetAccessRequests", options: {}, I: GetAccessRequestsRequest, O: GetAccessRequestsResponse },
     { name: "GetAccessRequest", options: {}, I: GetAccessRequestRequest, O: GetAccessRequestResponse },
     { name: "DeleteAccessRequest", options: {}, I: DeleteAccessRequestRequest, O: EmptyResponse },
